requirements.txt
google-genai
pypdf
streamlit
numpy
python-dotenv